+++ exporter/opentelemetry-exporter-otlp-proto-http/src/opentelemetry/exporter/otlp/proto/http/_log_exporter/__init__.py
@@ -146,7 +146,7 @@ class OTLPLogExporter(LogExporter):
 
             resp = self._export(serialized_data)
             # pylint: disable=no-else-return
-            if resp.status_code in (200, 202):
+            if resp.ok:
                 return LogExportResult.SUCCESS
             elif self._retryable(resp):
                 _logger.warning(